# API.yaml
openapi: 3.0.3
info:
  title: Printer Queue Generation API
  description: |
    API para geração de filas de impressão com empacotamento automático de imagens.
    
    ## Funcionalidades Principais
    
    - **Gestão de Tenants**: Multi-tenancy com isolamento de dados
    - **Gestão de Máquinas**: Configuração de máquinas de impressão por tenant
    - **Gestão de Assets**: Catálogo de imagens indexadas por SKU
    - **Gestão de Perfis de Dimensionamento**: Perfis para dimensionamento automático de imagens
    - **Gestão de Storage**: Configuração de storage (S3, Local, Dropbox)
    - **Processamento de Jobs**: Upload de picklists PDF, resolução de SKUs, empacotamento e geração de PDFs
    - **Image Packing**: Upload direto de imagens para empacotamento sem validação de SKU
    
    ## Autenticação
    
    A API usa multi-tenancy baseado em header `X-Tenant-ID`. Todos os endpoints (exceto `/healthz`) requerem este header.
    
    ## Status de Jobs
    
    - `queued`: Job criado, aguardando processamento
    - `processing`: Job sendo processado pelo worker
    - `needs_input`: Job precisa de resolução manual de SKUs
    - `completed`: Job processado com sucesso
    - `failed`: Job falhou durante processamento
    
    ## Modos de Processamento
    
    - `sequence`: Processa itens na ordem exata do picklist PDF
    - `optimize`: Otimiza layout para minimizar número de bases e desperdício
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
...
paths:
...
  /v1/assets:
    get:
      tags:
        - assets
      summary: List assets with pagination
      operationId: listAssets
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
    
    post:
      tags:
        - assets
      summary: Create asset manually
      operationId: createAsset
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCreate'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
  /v1/assets/reindex:
    post:
      tags:
        - assets
      summary: Trigger asset reindexation
      description: |
        Dispatches an asynchronous task to reindex all assets from tenant's storage.
        The task will:
        - List all files in storage
        - Extract SKU from filenames
        - Extract image metadata (dimensions, DPI, format)
        - Upsert assets in database
      operationId: reindexAssets
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetReindexRequest'
      responses:
        '202':
          description: Reindexation task started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetReindexResponse'
  /v1/assets/reindex/{task_id}:
    get:
      tags:
        - assets
      summary: Get reindexation task status
      operationId: getReindexStatus
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetReindexStatus'
  /v1/assets/search:
    post:
      tags:
        - assets
      summary: Search assets by SKU
      description: Fuzzy search for assets by SKU (normalized)
      operationId: searchAssets
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetSearchResponse'
...
components:
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant ID for multi-tenancy
      schema:
        type: integer
...
  schemas:
...
    AssetResponse:
      type: object
      properties:
        id:
          type: integer
        tenant_id:
          type: integer
        sku_normalized:
          type: string
        original_filename:
          type: string
        file_uri:
          type: string
        metadata_json:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AssetCreate:
      type: object
      required:
        - tenant_id
        - sku_normalized
        - original_filename
        - file_uri
      properties:
        tenant_id:
          type: integer
        sku_normalized:
          type: string
        original_filename:
          type: string
        file_uri:
          type: string
        metadata_json:
          type: string
          nullable: true
    AssetListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AssetResponse'
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
    AssetReindexRequest:
      type: object
      properties:
        force:
          type: boolean
          default: false
          description: Force reindex even if assets exist
    AssetReindexResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
        message:
          type: string
    AssetReindexStatus:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, started, success, failure]
        result:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
    AssetSearchRequest:
      type: object
      required:
        - sku
      properties:
        sku:
          type: string
        limit:
          type: integer
          default: 10
          maximum: 50
    AssetSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              asset_id:
                type: integer
              sku:
                type: string
              file_uri:
                type: string
              score:
                type: number
                format: float
...